+++ lib/mainImpl.dart
@@ -66,6 +66,7 @@ import 'package:rokwire_plugin/service/location_services.dart';
 import 'package:rokwire_plugin/service/app_navigation.dart';
 import 'package:rokwire_plugin/service/firebase_core.dart';
 import 'package:rokwire_plugin/service/firebase_crashlytics.dart';
+import 'package:rokwire_plugin/service/local_notifications.dart';
 import 'package:rokwire_plugin/service/service.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/app_livecycle.dart';
@@ -118,6 +119,7 @@ void mainImpl({ rokwire.ConfigEnvironment? configEnvironment }) async {
     Styles(),
     Analytics(),
     FirebaseMessaging(),
+    LocalNotifications(),
     Sports(),
     LiveStats(),
     RecentItems(),